Add audio chunker with two example devcards
The audio chunker can be placed in any audio pipeline. It will forward
input audio to its output as-is but will gather chunks with a specific
amount of samples on the side and deliver them via a callback. This
allows to slice up input audio into chunks of any size and eventually
encode these separately from each other.